Trac Git Hook
Post-receive hook for git allowing to control workflow of open tickets
in Trac issue tracker.
Using test #NUM in its commit, one can easily pass the ticket as
"Testing" or "Fixed on branch" (if the commit is on a feature branch).
When the feature is merged on the master branch, the script will
automagically reopen the ticket and set it to "Testing"